Add files via upload
This commentary reflects on the macroeconomic and financial outlook in 2009 when the global financial drew to a close and the recession officially ended. Specifically, I have discussed the Federal Reserve’s exit strategy, its normalization policy through the premature hike in the federal funds rate, and the financial market’s reaction via a drop in the 10-year Treasury yield. To that end, I conducted a structural break analysis (Chow test) to examine the impact of the former Chairman Ben Bernanke’s statements (regarding the ballooning debt and fiscal deficits) on the investors’ expectations. Gauging the monetization rates over time more closely, I finally reviewed the literature on the effectiveness of the stimulus package - American Recovery and Reinvestment Act, and how fiscal forward guidance can be considered as an optimal communication tool.